requirements.txt
# Dependências principais da app Assist-Decivil
streamlit>=1.33.0
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-docx>=0.8.11
PyPDF2>=3.0.1
beautifulsoup4>=4.12.2
pymupdf>=1.23.8
requests>=2.31.0
supabase>=2.4.0
python-dotenv>=1.0.1
tqdm>=4.66.4
